Refactor tests: extract shared helpers to tests/helpers.py
